vector_server/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
cachetools>=5.5,<6
httpx>=0.28,<1
cryptography>=44,<45
reportlab>=4.3,<5
Pillow>=11,<12
Jinja2>=3.1,<4

# dev / test
pytest>=8,<10
pytest-asyncio>=0.24,<2
